+++ src/wo_lang_ast_builder.hpp
@@ -1227,7 +1227,7 @@ namespace wo
                         }
 
                         wchar_t out_str = u8stridx(from->get_constant_value().string->c_str(),
-                            index->get_constant_value().integer);
+                            (size_t)index->get_constant_value().integer);
                         if (out_str == 0 && u8strlen(from->get_constant_value().string->c_str())
                             <= (size_t)index->get_constant_value().integer)
                             lex->lang_error(lexer::errorlevel::error, index, WO_ERR_INDEX_OUT_OF_RANGE);
